[rhi] simplified resource layout, everything is tracked by the command list now, this means both textures are swapchains are less error prone now

+++ runtime/RHI/RHI_Definitions.h
@@ -460,7 +460,7 @@ namespace spartan
     const uint8_t  rhi_max_constant_buffer_count = 8;
     const uint32_t rhi_max_array_size            = 16384;
     const uint32_t rhi_max_descriptor_set_count  = 512;
-    const uint8_t  rhi_max_mip_count             = 13;
+    const uint32_t  rhi_max_mip_count            = 13;
     const uint32_t rhi_all_mips                  = std::numeric_limits<uint32_t>::max();
     const uint32_t rhi_dynamic_offset_empty      = std::numeric_limits<uint32_t>::max();
     const uint32_t rhi_max_buffer_update_size    = 65536; // vkCmdUpdateBuffer has a limit of 65536 bytes
